+++ include/LIEF/ELF/ParserConfig.hpp
@@ -25,7 +25,7 @@ namespace ELF {
 struct LIEF_API ParserConfig {
   /** Methods that can be used by the LIEF::ELF::Parser
       to count the number of dynamic symbols */
-  enum class DYNSYM_COUNT  {
+  enum class DYNSYM_COUNT {
     AUTO        = 0, /**< Automatic detection */
     SECTION     = 1, /**< Count based on sections (not very reliable) */
     HASH        = 2, /**< Count based on hash table (reliable) */
@@ -48,6 +48,11 @@ struct LIEF_API ParserConfig {
 
   /** The method used to count the number of dynamic symbols */
   DYNSYM_COUNT count_mtd = DYNSYM_COUNT::AUTO;
+
+  /// Memory page size if the binary uses a non-standard value.
+  ///
+  /// For instance, SPARCV9 binary can use page size from 0x2000 to 0x100000.
+  uint64_t page_size = 0;
 };
 
 }