+++ tests/abstract/test_abstract.py
@@ -119,3 +119,14 @@ def test_entropy():
 def test_issue_1217():
     elf = lief.ELF.parse(get_sample("ELF/bitcoin_ppc_be"))
     assert elf.abstract.header.architecture == lief.Header.ARCHITECTURES.PPC64
+
+def test_pagesize():
+    assert lief.parse(get_sample('ELF/ELF64_x86-64_library_libadd.so')).page_size == 0x1000
+    assert lief.parse(get_sample('PE/win11_arm64x_api-ms-win-security-base-l1-1-0.dll')).page_size == 0x1000
+    assert lief.parse(get_sample('MachO/MachO64_AArch64_weak-sym-fc.bin')).page_size == 0x4000
+
+    config = lief.ELF.ParserConfig()
+    config.page_size = 0xdeadc0de
+
+    assert lief.ELF.parse(get_sample('ELF/bitcoin_ppc_be'), config).page_size == config.page_size
+